move cache to own module
